+++ frontend/src/app/page.tsx
@@ -1,28 +1,23 @@
 import React from "react";
-import { auth, currentUser } from "@clerk/nextjs";
-import ChooseRoleDialog from "@/components/ChooseRoleDialog";
+import Link from "next/link";
 
 
 export default async function Home() {
   // Get the userId from auth() -- if null, the user is not logged in
-  const { userId } = auth();
-
-  const theUser = await currentUser();
-  // Get the User object when you need access to the user's information
-  const user = {
-    userId,
-    firstName: theUser?.firstName ?? '',
-    lastName: theUser?.lastName ?? '',
-    emailAddress: theUser?.emailAddresses?.[0]?.emailAddress ?? null,
-    phoneNumber: theUser?.phoneNumbers?.[0]?.phoneNumber ?? null,
-    imageUrl: theUser?.imageUrl ?? null,
-    authenticationMethod: theUser?.externalAccounts[0]?.verification?.strategy,
-  };
-
   return (
     <>
       {/* 現在預設會直接打開 ChooseRoleDialog */}
-      <ChooseRoleDialog props={user}/>
+      {/* <ChooseRoleDialog props={user}/> */}
+      <div className="flex flex-col min-h-screen">
+        <main className="flex flex-col items-center p-4 pt-20 mx-auto w-[95%] grow">
+          <h1 className="text-3xl font-bold text-center">
+            Welcome to Foody!
+          </h1>
+          <Link href="/login">
+            <span className="text-blue-500 underline">Go to Restaurant</span>
+          </Link>
+        </main>
+      </div>
     </>
   );
 }
